+++ data_gen/data_gen/simple.py
@@ -1,0 +1,29 @@
+import json
+import dataclasses
+from json import JSONEncoder
+from mygfa import mygfa
+
+
+class GenericSimpleEncoder(JSONEncoder):
+    def default(self, o):
+        if isinstance(o, mygfa.Path):
+            items = str(o).split("\t")
+            return {"segments": items[2], "overlaps": items[3]}
+        if isinstance(o, mygfa.Link):
+            return {
+                "from": o.from_.name,
+                "from_orient": o.from_.orientation,
+                "to": o.to.name,
+                "to_orient": o.to.orientation,
+                "overlap": str(o.overlap),
+            }
+        if isinstance(o, mygfa.Segment) or isinstance(o, mygfa.Alignment):
+            return dataclasses.asdict(o)
+
+
+def simple(graph):
+    """Prints a "wholesale dump" JSON representation of `graph`"""
+    print(json.dumps(graph.headers, indent=4))
+    print(json.dumps(graph.segments, indent=4, cls=GenericSimpleEncoder))
+    print(json.dumps(graph.links, indent=4, cls=GenericSimpleEncoder))
+    print(json.dumps(graph.paths, indent=4, cls=GenericSimpleEncoder))
